fix: failed to detect the execution user

+++ main.sh
@@ -36,6 +36,10 @@ function ERROR() {
 function WARN() {
     echo -e "${WARN} ${1}"
 }
+if [[ $EUID -ne 0 ]]; then
+    ERROR '此脚本必须以 root 身份运行！'
+    exit 1
+fi
 if [ -f /tmp/xiaoya_install.sh ]; then
     rm -rf /tmp/xiaoya_install.sh
 fi
@@ -43,6 +47,7 @@ if [ -n "${XIAOYA_BRANCH}" ]; then
     if ! curl -sL "https://fastly.jsdelivr.net/gh/DDS-Derek/xiaoya-alist@${XIAOYA_BRANCH}/all_in_one.sh" -o /tmp/xiaoya_install.sh; then
         if ! curl -sL "https://raw.githubusercontent.com/DDS-Derek/xiaoya-alist/${XIAOYA_BRANCH}/all_in_one.sh" -o /tmp/xiaoya_install.sh; then
             ERROR "脚本获取失败！"
+            exit 1
         else
             bash /tmp/xiaoya_install.sh $@
         fi
@@ -54,6 +59,7 @@ else
         if ! curl -sL https://fastly.jsdelivr.net/gh/DDS-Derek/xiaoya-alist@latest/all_in_one.sh -o /tmp/xiaoya_install.sh; then
             if ! curl -sL https://raw.githubusercontent.com/DDS-Derek/xiaoya-alist/master/all_in_one.sh -o /tmp/xiaoya_install.sh; then
                 ERROR "脚本获取失败！"
+                exit 1
             else
                 bash /tmp/xiaoya_install.sh $@
             fi
